Face-Recognition with Emotion Detection/face_recognization with Emotion/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
pymongo==4.6.1
opencv-python==4.8.0.76
face-recognition==1.3.0
numpy==1.24.4
tensorflow==2.13.0
scikit-learn==1.3.2
Pillow==10.2.0
